Added better support for writing acid and Tracktion chunks in WavAudioFormat.

+++ modules/juce_audio_formats/format/juce_AudioFormatReader.cpp
@@ -176,7 +176,7 @@ void AudioFormatReader::read (AudioSampleBuffer* buffer,
 void AudioFormatReader::readMaxLevels (int64 startSampleInFile, int64 numSamples,
                                        Range<float>* const results, const int channelsToRead)
 {
-    jassert (channelsToRead > 0 && channelsToRead <= numChannels);
+    jassert (channelsToRead > 0 && channelsToRead <= (int) numChannels);
 
     if (numSamples <= 0)
     {
